# backend/internal/http/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Mappening API
  version: 0.1.0
  description: API backend de Mappening.
servers:
  - url: http://localhost:8080
    description: Backend local
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Connecte un compte.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session creee.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/auth/register/user:
    post:
      tags: [Auth]
      security: []
      summary: Cree un compte utilisateur.
      responses:
        "201":
          description: Compte utilisateur cree.
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/register/organization:
    post:
      tags: [Auth]
      security: []
      summary: Cree un compte organisation.
      responses:
        "201":
          description: Compte organisation cree.
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Retourne le compte courant.
      responses:
        "200":
          description: Compte courant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/auth/check-role/{role}:
    get:
      tags: [Auth]
      summary: Verifie le role du compte courant.
      parameters:
        - $ref: "#/components/parameters/RolePath"
      responses:
        "200":
          description: Resultat du controle.
  /api/auth/check-account-type/{accountType}:
    get:
      tags: [Auth]
      summary: Verifie le type de compte courant.
      parameters:
        - name: accountType
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resultat du controle.
...
  /api/admin/users/{userID}:
    patch:
      tags: [Admin]
      summary: Met a jour un utilisateur.
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Utilisateur mis a jour.
    delete:
      tags: [Admin]
      summary: Supprime un utilisateur.
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: Utilisateur supprime.
  /api/admin/users/{userID}/reset-password:
    post:
      tags: [Admin]
      summary: Reinitialise le mot de passe d'un utilisateur.
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: Mot de passe reinitialise.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: integer
...
    RolePath:
      name: role
      in: path
      required: true
      schema:
        type: string
        enum: [admin, moderator, organization, user]
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        role:
          type: string
        is_active:
          type: boolean
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Requete invalide.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentification requise ou invalide.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...